Extracted FetchData and FetchFaveData into separate classes
